% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benton}
\alias{benton}
\title{Trip records data for Benton County, OR}
\format{
\subsection{\code{benton}}{

A data.table with 364,091 rows and 3 columns:
\describe{
\item{GEOID}{block group ID}
\item{start_time}{trip start time (in seconds)}
\item{end_time}{trip end time (in seconds)}
...
}
}
}
\source{
Acquired using Renaissance Planning's license to Replica; for more
information, see \url{https://www.replicahq.com/}
}
\usage{
benton
}
\description{
Simulated trip start and end times for all trips originating in Benton
County, OR on a typical Thursday in Q4 of 2021. Trips may be grouped by the
origin block group (GEOID).
}
\keyword{datasets}
