% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_trips_curves.R
\name{combine_completed_trips_by_minute}
\alias{combine_completed_trips_by_minute}
\title{Combine multiple completed trips by minute vectors into a single vector}
\usage{
combine_completed_trips_by_minute(...)
}
\arguments{
\item{...}{Any number of completed trips by minute vectors}
}
\value{
A vector of completed trips by minute, as totaled from the input
vectors
}
\description{
This function combines any number of complete trip by minute vectors into a
single completed trips by minute vector by adding completed trips at each
minute
}
