% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_trips_curves.R
\name{complete_trips_curve}
\alias{complete_trips_curve}
\title{Convert a completed trips by minute vector into a completed trips curve}
\usage{
complete_trips_curve(x, percent = FALSE)
}
\arguments{
\item{x}{A completed trips by minute vector}

\item{percent}{A boolean indicating whether the curve should be constructed
as total trips completed (\code{FALSE}) or share of trips completed (\code{TRUE}).
Setting \code{percent=TRUE} allows direct comparisons to be made across groupings
that have different magnitudes of completed trips}
}
\value{
A function defining the completed trips curve for the input completed
trips by minute. It will span a 2-day period to guarantee continuity at
12:00am
}
\description{
Convert a completed trips by minute vector into a completed trips curve
}
