% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_trips_curves.R
\name{completed_trips_by_minute}
\alias{completed_trips_by_minute}
\title{Calculate completed trips by minute from a set of trip start and end times}
\usage{
completed_trips_by_minute(start_times, end_times)
}
\arguments{
\item{start_times}{Vector of trip start times, in integer seconds; should
align with indices of \code{end_times}}

\item{end_times}{Vector of trip end times, in integer seconds; should align
with indices of \code{start_times}}
}
\value{
A vector of completed trips for each minute in a 24 hour day,
beginning and ending at 12:00am (for a total of 1441 entries)
}
\description{
Calculate completed trips by minute from a set of trip start and end times
}
