% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{efficiency}
\alias{efficiency}
\title{Calculate trip efficiency}
\usage{
efficiency(time_lowers, time_uppers, complete_trips_curve)
}
\arguments{
\item{time_lowers}{Lower bound for time intervals; should align with indices
of \code{time_uppers}.}

\item{time_uppers}{Upper bound for time intervals; should align with indices
of \code{time_lowers}}

\item{complete_trips_curve}{The complete trips curve on which to assess
efficiency}
}
\value{
A vector of efficiencies for each interval implied by \code{time_lowers}
and \code{time_uppers}
}
\description{
"Efficiency" is defined as the expected number of completed trips observed
(or share of completed trips observed) at a random point in a specified
interval. That is to say, for a random time \eqn{t: t1 <= t <= t2}, how many
completed trips (or what share of total trips completed) would we expect to
have observed?
}
