% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspacman-package.R
\docType{package}
\name{mspacman-package}
\alias{mspacman}
\alias{mspacman-package}
\title{mspacman: Multimodal Systems Productivity: Analytic and Computational Methods, Approaches, and Needs}
\description{
Uses trip start- and end-time data to build complete trip curves for a collection of trips. From these curves, metrics such as productivity and efficiency can be calculated, and comparing these metrics across collections of trips allows for quick hit insight on variations in observed travel.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/renaissanceplanning/mspacman}
}

}
\author{
\strong{Maintainer}: Aaron Weinstock \email{aweinstock@citiesthatwork.com}

}
\keyword{internal}
