% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_complete_trips_curve}
\alias{plot_complete_trips_curve}
\title{Plot a complete trips curve}
\usage{
plot_complete_trips_curve(complete_trips_curve)
}
\arguments{
\item{complete_trips_curve}{A complete trips curve}
}
\value{
a \code{ggplot2} plot of the complete trips curve
}
\description{
Plot a complete trips curve
}
