% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_efficiency}
\alias{plot_efficiency}
\title{Plot efficiency for a complete trips curve}
\usage{
plot_efficiency(complete_trips_curve, window = 1, smooth = TRUE)
}
\arguments{
\item{complete_trips_curve}{A complete trips curve}

\item{window}{The size of the window (in minutes) over which efficiency
should be calculated}

\item{smooth}{Should the curve be smoothed using \code{ggplot::geom_smooth}?}
}
\value{
a \code{ggplot2} plot of efficiency
}
\description{
Plot efficiency for a complete trips curve
}
