% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_productivity}
\alias{plot_productivity}
\title{Plot productivity for a complete trips curve}
\usage{
plot_productivity(complete_trips_curve, smooth = TRUE)
}
\arguments{
\item{complete_trips_curve}{A complete trips curve}

\item{smooth}{Should the curve be smoothed using \code{ggplot::geom_smooth}?}
}
\value{
a \code{ggplot2} plot of productivity
}
\description{
Plot productivity for a complete trips curve
}
