% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{productivity}
\alias{productivity}
\title{Calculate trip productivity}
\usage{
productivity(times, complete_trips_curve)
}
\arguments{
\item{times}{A vector of times specified in minutes (and therefore between
0 and 1440, inclusive)}

\item{complete_trips_curve}{The complete trips curve on which to assess
productivity}
}
\value{
A vector of productivities for each time in \code{times}
}
\description{
"Productivity" is defined as the first derivative of the complete trips
curve at time \eqn{t}
}
